[js] Factor out future to its own file / interface.
This is in preparation towards a more generic Notification/Request
model for the servers.
